Compare full HDF5 output files instead of pickled hashes for bit tests as different compilers can give different results in terms of floating point precision. This involved making sure that the PDF sampling code doesn't use tiny(), which is compiler-dependent.

+++ src/grid/type_grid_octree.f90
@@ -10,8 +10,8 @@ module type_grid
   ! Define refinable cell
   public :: cell
   type cell
-     real(8) :: x,y,z
-     real(8) :: dx,dy,dz
+     real(dp) :: x,y,z
+     real(dp) :: dx,dy,dz
      logical :: refined = .false.
      integer,allocatable,dimension(:) :: children
      integer :: parent, parent_subcell
